add authorization for CRUD requests

+++ server/controllers/top5list-controller.js
@@ -35,6 +35,15 @@ createTop5List = (req, res) => {
 }
 
 updateTop5List = async (req, res) => {
+    const user_id = req.userId;
+    let user_email = "";
+    await User.findById({_id: user_id}, (err, user) => {
+        if(err) {
+            return res.status(401).json({success:false, error:err});
+        }
+        user_email = user.email;
+    })
+
     const body = req.body;
     console.log("updateTop5List: " + JSON.stringify(body));
     if (!body) {
@@ -51,6 +60,9 @@ updateTop5List = async (req, res) => {
                 err,
                 message: 'Top 5 List not found!',
             })
+        }
+        if(user_email !== top5List.owner){
+            return res.status(403).json({success: false, error: "You are not authorized to edit this list." });
         }
 
         top5List.name = body.name
@@ -77,6 +89,15 @@ updateTop5List = async (req, res) => {
 }
 
 deleteTop5List = async (req, res) => {
+    const user_id = req.userId;
+    let user_email = "";
+    await User.findById({_id: user_id}, (err, user) => {
+        if(err) {
+            return res.status(401).json({success:false, error:err});
+        }
+        user_email = user.email;
+    })
+
     Top5List.findById({ _id: req.params.id }, (err, top5List) => {
         if (err) {
             return res.status(404).json({
@@ -84,6 +105,11 @@ deleteTop5List = async (req, res) => {
                 message: 'Top 5 List not found!',
             })
         }
+
+        if(user_email !== top5List.owner){
+            return res.status(403).json({success: false, error: "You are not authorized to delete this list."});
+        }
+
         Top5List.findOneAndDelete({ _id: req.params.id }, () => {
             return res.status(200).json({ success: true, data: top5List })
         }).catch(err => console.log(err))
@@ -105,7 +131,7 @@ getTop5ListById = async (req, res) => {
             return res.status(400).json({ success: false, error: err });
         }
         if(user_email !== list.owner){
-            return res.status(403).json({success: false, error: err });
+            return res.status(403).json({success: false, error: "You are not authorized to edit this list." });
         }
         return res.status(200).json({ success: true, top5List: list })
     }).catch(err => console.log(err))
